Bugfix for CREATE_PROCESS
Due to the fact that prcs_info becomes malloced, it might be that initial data is not 0. However, `CREATE_PROCESS` tries to search for a `prcs_info[idx].id` which shall be 0. In the worst case, this won't happen. Since idx is also not checked against MAX_PRCS, the idx can be way out of the malloced reachen.